regenerated file: downloads/index.md
The names of the downloadable files are now rendered in the HTML as
hyperlinks.

This change required making a trade-off between readability and clutter in the
plain markdown version of the file. Since we want the markdown to remain
readable, we do not want to list the filename twice in each row of the data
table (once for the link label and once for the link target). To achieve that,
we need to include a separate link definition table somewhere in the markdown
file; this changeset places it near the bottom of the file.

In the data table, hyperlinks are created using the file name as the index
key. Those keys refer to entries in the link definitions table. The markdown
processor uses them to create the hyperlink href targets when translating the
the file to HTML.

The alternative would be to omit the link definition table and instead
indicate all of the link information inline, but that would be unreadable in
the markdown; the links would need to be created in the form:

    [link label text][link target url]

Compare:

    | [ads-clock-0.1.0.tar.gz.SHA-384][ads-clock-0.1.0.tar.gz.SHA-384]  | `2016-06-12 22:22:51` |          `121` |          `121` |

with the more readable:

    | [ads-clock-0.1.0.tar.gz.SHA-384][]  | `2016-06-12 22:22:51` |          `121` |          `121` |

The trade-off for the improved readability of the second form, of course, is
that we need the link definition table somewhere in our markdown document,
which adds clutter. Note that the link definition table will not appear in the
rendered HTML.