regenerated file: downloads/index.md
Make 'ads-clock' portion of "Index of /ads-clock/downloads/" text a hyperlink
to the parent directory, which will be the index page of the project. This
gives the user an easy way to find the project page if they happen to stuble
directly onto the downloads page for some reason (as opposed to landing on the
downloads page by coming from the project page).