regenerated file: downloads/index.md
I noticed that timestamps reported for the downloadable files in the
downloads/index.md file reflected the current time on my system rather than
the time those files were comitted to the git repo. I used the
'git-set-file-times' script (from the Debian 'rsync-3.1.1-3' package) to reset
them, and will need to do this going forward when the file is regenerated.

So the cookbook is now:

    $ git-set-file-times && ( cd downloads && ads-md-table-for-downloads > index.md )